+++ views/generator/type/clone.php
@@ -1,5 +1,5 @@
 /**
- * <?php echo ucfirst($type), ' ', $name ?>, cloned from <?php echo $source ?>.
+ * <?php echo ucfirst($type), ' ', $name, ', cloned from ', $source ?>.
  *
  * @package    <?php echo $package ?> 
  * @category   <?php echo $category ?> 
@@ -8,12 +8,18 @@
  * @license    <?php echo $license ?> 
  */
 <?php 
-	echo ( ! empty($modifiers) ? ($modifiers.' '.$type) : $type);
-	echo ' ', $name;
+	if ( ! empty($modifiers)) {echo $modifiers, ' ';}
+	echo $type, ' ', $name;
 	if ( ! empty($extends)) {echo ' extends ', $extends;}
 	if ( ! empty($implements)) {echo ' implements ', $implements;}
 	if ( ! empty($blank)): echo ' {}'; else : ?> 
 {
+<?php if ( ! empty($traits)): foreach ($traits as $trait): ?>
+	// Trait: <?php echo $trait ?> 
+	<?php echo 'use ', $trait, ';' ?> 
+
+<?php endforeach; ?>
+<?php endif; ?>
 <?php if ( ! empty($constants)) foreach ($constants as $constant => $c): ?>
 	<?php echo $c['comment'] ?> 
 	<?php echo $c['declaration'] ?>;